fixed #82 by re-adding the witting the of jsData from content.js in warcGenerator.js
penitential fix for #17 and #78 by replacing '\n' with '\r\n' in WARCEntryCreator.touchUpInitURIHeaders in warcGenerator.js

fixed #79 by adding listener to onSendHeaders. Avoid duplicates by keeping track per URL which headers have been added so far

refactored content.js per #90 and to ensure all of a pages resources are in the warc file
which facilitated the fix for #82

fixed #80 the metadata information for CSS is no longer E =EMBED_MIC it is now E link/@href

fixed #81 the reason why the script tag links were missing from the warcmetadata
was because the code checking for its existence was looking for the HREF attribute
not src attribute. Not rhetorical this error was simply due to an copy pasta error.
Was the exact code for outlinks as CSS and have I ever made my share of these :)