Catch partial URIs without any relativity indicator
Hacker News, for instance, references a JS file with just the file
name without any relative indicator (./ or ../), any scheme, or
really anything that would allow it to be parsed as a full URI
with native functionality. This adds a check for this TypeError
and attributes the current hostname/scheme to the partial.

Closes #128